fixed bug.

+++ main.js
@@ -1,4 +1,39 @@
 "use strict";
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
+var __generator = (this && this.__generator) || function (thisArg, body) {
+    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
+    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
+    function verb(n) { return function (v) { return step([n, v]); }; }
+    function step(op) {
+        if (f) throw new TypeError("Generator is already executing.");
+        while (_) try {
+            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
+            if (y = 0, t) op = [op[0] & 2, t.value];
+            switch (op[0]) {
+                case 0: case 1: t = op; break;
+                case 4: _.label++; return { value: op[1], done: false };
+                case 5: _.label++; y = op[1]; op = [0]; continue;
+                case 7: op = _.ops.pop(); _.trys.pop(); continue;
+                default:
+                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
+                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
+                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
+                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
+                    if (t[2]) _.ops.pop();
+                    _.trys.pop(); continue;
+            }
+            op = body.call(thisArg, _);
+        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
+        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
+    }
+};
 exports.__esModule = true;
 var _a = require('electron'), app = _a.app, BrowserWindow = _a.BrowserWindow, Tray = _a.Tray, ipcMain = _a.ipcMain, shell = _a.shell, Menu = _a.Menu;
 var spawn = require('child_process').spawn;
@@ -8,6 +43,7 @@ var HTTPPlus = require('./HTTPPlus');
 var fs = require('fs');
 var queue = require('queue');
 var dateFormat = require('dateformat');
+var download = require('download');
 var isdelts = true;
 var mainWindow = null;
 var playerWindow = null;
@@ -31,6 +67,7 @@ function createWindow() {
     mainWindow.setMenu(null);
     // 加载index.html文件
     mainWindow.loadFile(path.join(__dirname, 'index.html'));
+    //mainWindow.openDevTools();
     // 当 window 被关闭，这个事件会被触发。
     mainWindow.on('closed', function () {
         // 取消引用 window 对象，如果你的应用支持多窗口的话，
@@ -114,6 +151,11 @@ app.on('activate', function () {
         createWindow();
     }
 });
+ipcMain.on("hide-windows", function () {
+    if (mainWindow != null) {
+        mainWindow.hide();
+    }
+});
 ipcMain.on('get-all-videos', function (event, arg) {
     event.sender.send('get-all-videos-reply', configVideos);
 });
@@ -151,6 +193,7 @@ ipcMain.on('task-add', function (event, arg) {
     });
 });
 function startDownload(url, parser) {
+    var _this = this;
     var partent_uri = url.replace(/([^\/]*\?.*$)|([^\/]*$)/g, '');
     var id = new Date().getTime();
     var dir = path.join(app.getAppPath().replace(/resources\\app.asar$/g, ""), 'download/' + id);
@@ -173,25 +216,60 @@ function startDownload(url, parser) {
     mainWindow.webContents.send('task-notify-create', video);
     parser.manifest.segments.forEach(function (segment) {
         console.log("1 " + segment.uri);
-        q.push(function (cb) {
-            var uri_ts = '';
-            if (/^http.*/.test(segment.uri)) {
-                uri_ts = segment.uri;
-            }
-            else {
-                uri_ts = partent_uri + segment.uri;
-            }
-            console.log("2 " + segment.uri);
-            var filpath = path.join(dir, segment.uri.replace(/(^.*\/)|(\?.*$)/g, ''));
-            HTTPPlus.downloadFileSync(uri_ts, filpath, 'GET', null, null, function (res_data, error) {
-                console.log("3 " + segment.uri);
-                count_downloaded = count_downloaded + 1;
-                video.segment_downloaded = count_downloaded;
-                video.status = "\u4E0B\u8F7D\u4E2D..." + count_downloaded + "/" + count_seg;
-                mainWindow.webContents.send('task-notify-update', video);
-                cb(error, res_data);
+        q.push(function (cb) { return __awaiter(_this, void 0, void 0, function () {
+            var uri_ts, filename, filpath, index;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        uri_ts = '';
+                        if (/^http.*/.test(segment.uri)) {
+                            uri_ts = segment.uri;
+                        }
+                        else {
+                            uri_ts = partent_uri + segment.uri;
+                        }
+                        console.log("2 " + segment.uri);
+                        filename = segment.uri.replace(/(^.*\/)|(\?.*$)/g, '');
+                        filpath = path.join(dir, filename);
+                        index = 0;
+                        _a.label = 1;
+                    case 1:
+                        if (!(index < 3)) return [3 /*break*/, 5];
+                        if (!!fs.existsSync(filpath)) return [3 /*break*/, 3];
+                        // 下载的时候使用.dl后缀的文件名，下载完成后重命名
+                        return [4 /*yield*/, download(uri_ts, dir, { filename: filename + ".dl" }).then(function () {
+                                fs.renameSync(filpath + ".dl", filpath);
+                            })["catch"](function () {
+                                try {
+                                    fs.unlinkSync(filpath + ".dl");
+                                }
+                                catch (derror) {
+                                    console.log(derror);
+                                }
+                            })];
+                    case 2:
+                        // 下载的时候使用.dl后缀的文件名，下载完成后重命名
+                        _a.sent();
+                        _a.label = 3;
+                    case 3:
+                        if (fs.existsSync(filpath)) {
+                            console.log("3 " + segment.uri);
+                            count_downloaded = count_downloaded + 1;
+                            video.segment_downloaded = count_downloaded;
+                            video.status = "\u4E0B\u8F7D\u4E2D..." + count_downloaded + "/" + count_seg;
+                            mainWindow.webContents.send('task-notify-update', video);
+                            return [3 /*break*/, 5];
+                        }
+                        _a.label = 4;
+                    case 4:
+                        index++;
+                        return [3 /*break*/, 1];
+                    case 5:
+                        cb(null, "success");
+                        return [2 /*return*/];
+                }
             });
-        });
+        }); });
     });
     q.start(function () {
         console.log('download success');
@@ -204,15 +282,20 @@ function startDownload(url, parser) {
             filesegments.push(filpath);
         });
         fs.writeFileSync(path.join(dir, 'index.txt'), indexData);
-        var p = spawn(path.join(app.getAppPath().replace(/resources\\app.asar$/g, ""), "ffmpeg.exe"), ["-f", "concat", "-safe", "0", "-i", "" + path.join(dir, 'index.txt'), "-c", "copy", "" + path.join(dir, id + '.mp4')]);
+        var p = spawn(path.join(app.getAppPath().replace(/resources\\app.asar$/g, ""), "ffmpeg"), ["-f", "concat", "-safe", "0", "-i", "" + path.join(dir, 'index.txt'), "-c", "copy", "" + path.join(dir, id + '.mp4')]);
         p.on("close", function () {
             video.videopath = path.join(dir, id + '.mp4');
             video.status = "已完成";
             mainWindow.webContents.send('task-notify-end', video);
             if (isdelts) {
-                fs.unlinkSync(path.join(dir, 'index.txt'));
+                var index_path = path.join(dir, 'index.txt');
+                if (fs.existsSync(index_path)) {
+                    fs.unlinkSync(index_path);
+                }
                 filesegments.forEach(function (fileseg) {
-                    fs.unlinkSync(fileseg);
+                    if (fs.existsSync(fileseg)) {
+                        fs.unlinkSync(fileseg);
+                    }
                 });
             }
             configVideos.push(video);
